+++ engine/flink/components/base-tests/src/test/scala/pl/touk/nussknacker/engine/flink/table/source/TableSourceDataGenerationTest.scala
@@ -24,7 +24,6 @@ class TableSourceDataGenerationTest
     .create(
       sqlStatements = Some(SqlStatementReader.readSql(SimpleTable.sqlStatement)),
       catalogConfigurationOpt = None,
-      defaultDbName = None
     )
     .validValue